+++ src/app/generator/test-data/param-generator/golden-files/amazon-workspaces.json
