+++ src/app/generator/test-data/param-generator/golden-files/auckland-museum.json
