+++ src/app/generator/test-data/param-generator/golden-files/aws-autoscaling.json
