+++ src/app/generator/test-data/param-generator/golden-files/azure-automation-management.json
