+++ src/app/generator/test-data/param-generator/golden-files/azure-open-id-connect.json
