+++ src/app/generator/test-data/param-generator/golden-files/cisco-meraki.json
