+++ src/app/generator/test-data/param-generator/golden-files/docker-dvp-data.json
