+++ src/app/generator/test-data/param-generator/golden-files/docker-hub.json
