+++ src/app/generator/test-data/param-generator/golden-files/ebay-finances.json
