+++ src/app/generator/test-data/param-generator/golden-files/gitlab.json
