+++ src/app/generator/test-data/param-generator/golden-files/google-adsense.json
