+++ src/app/generator/test-data/param-generator/golden-files/google-analytics-hub.json
