+++ src/app/generator/test-data/param-generator/golden-files/here-maps-positioning.json
