+++ src/app/generator/test-data/param-generator/golden-files/intuit-mailchimp.json
