+++ src/app/generator/test-data/param-generator/golden-files/medium.json
