+++ src/app/generator/test-data/param-generator/golden-files/mercedez-benz-car-configurator.json
