+++ src/app/generator/test-data/param-generator/golden-files/microsoft-workload-monitor.json
