+++ src/app/generator/test-data/param-generator/golden-files/snyk.json
